split downloader into separate crate
