+++ crates/plotinator-download/src/endpoint.rs
@@ -1,0 +1,20 @@
+#[derive(Debug, Clone, Copy, strum_macros::Display)]
+pub enum Endpoint {
+    #[strum(serialize = "/api/download/latest")]
+    DownloadLatestData,
+    #[strum(serialize = "/api/download/today")]
+    DownloadTodaysData,
+}
+
+#[cfg(test)]
+mod tests {
+    use super::*;
+    use pretty_assertions::assert_str_eq;
+
+    #[test]
+    fn test_serialize_endpoint() {
+        let endpoint = Endpoint::DownloadLatestData;
+
+        assert_str_eq!(endpoint.to_string(), "/api/download/latest");
+    }
+}
