+++ crates/plotinator-download/src/lib.rs
@@ -1,0 +1,15 @@
+pub enum DownloadMessage {
+    Success {
+        filename: String,
+    },
+    Error(String),
+    Progress {
+        downloaded_bytes: u64,
+        total_bytes: u64,
+    },
+    Finished,
+}
+
+pub(crate) mod downloader;
+pub mod endpoint;
+pub mod manager;
